+++ frontend/src/hooks/useTable.tsx
@@ -5,18 +5,12 @@ import { type TableData, handleSearchParams } from '@/components';
 import { AxiosPrivate } from '../utils';
 import { v4 as uuidv4 } from 'uuid';
 import { truncatePageRange } from './utils';
-import type {
-  AvailabilityInterface,
-  DashboardFilters,
-  Personnel,
-} from '@/pages/dashboard';
+import type { DashboardFilters, Personnel } from '@/pages/dashboard';
 import { DashboardColumns } from '@/pages/dashboard';
-import { tableClass } from '@/components/table/classes';
 import { useDebounce } from './useDebounce';
-import { AvailabilityType, AvailabilityTypeName, ExperienceName } from '@/common';
 import { useError } from './useError';
-import { differenceInDays } from 'date-fns';
 import type { DateRange } from 'react-day-picker';
+import { calculatePages, renderCells } from './helpers';
 
 export const useTable = () => {
   const { handleError } = useError();
@@ -42,7 +36,7 @@ export const useTable = () => {
       to: undefined,
     },
   });
-  const [showFunctionColumn, setShowFunctionColumn] = useState<boolean>(false);
+
   const [defaultDebounceValue, setDefaultDebounceValue] = useState(100);
   const [searchParamsUrl] = useSearchParams(encodeURI('?page=1&rows=25'));
   const debouncedValue = useDebounce<{ [key: string]: unknown }>(
@@ -50,49 +44,6 @@ export const useTable = () => {
     defaultDebounceValue,
   );
 
-  const calculatePages = (totalPages: number): number[] => {
-    const range = [];
-
-    // create the pages array for the footer
-    for (let i = 1; i <= totalPages; i++) {
-      range.push(i);
-    }
-    return range;
-  };
-  /**
-   * If availability  status is set, then calculate the partial or full availability
-   * @param availability
-   * @returns
-   */
-  const getAvailabilityValue = (availability: AvailabilityInterface[]) => {
-    const totalAvailableDays = availability.filter(
-      (itm) =>
-        itm.availabilityType ===
-        AvailabilityType[
-          filterValues?.availabilityType as keyof typeof AvailabilityType
-        ],
-    ).length;
-
-    const totalDaysSearched =
-      differenceInDays(
-        filterValues.availabilityDates.to ?? new Date(),
-        filterValues.availabilityDates.from ?? new Date(),
-      ) + 1;
-
-    if (totalDaysSearched === 1 || totalDaysSearched === totalAvailableDays) {
-      return AvailabilityTypeName[
-        filterValues?.availabilityType as keyof typeof AvailabilityType
-      ];
-    }
-
-    if (totalDaysSearched > 1 && totalAvailableDays >= 1) {
-      return AvailabilityTypeName.PARTIAL;
-    }
-
-    return AvailabilityTypeName[
-      filterValues?.availabilityType as keyof typeof AvailabilityType
-    ];
-  };
   useEffect(() => {
     (async () => {
       handleSearchParams(searchParamsUrl, filterValues);
@@ -106,120 +57,17 @@ export const useTable = () => {
         const totalPages = Math.ceil(count / rowsPerPage);
         const pageRange = calculatePages(Math.ceil(totalPages));
         const currentPage = filterValues?.currentPage ?? 1;
-        filterValues.function
-          ? setShowFunctionColumn(true)
-          : setShowFunctionColumn(false);
 
-        personnel &&
-          setTableData({
-            totalPages,
-            pageRange: truncatePageRange(totalPages, currentPage, pageRange),
-            totalRows: count,
-            rows: personnel.map(
-              ({
-                id,
-                status,
-                firstName,
-                lastName,
-                experiences,
-                region,
-                workLocation,
-                willingToTravel,
-                remoteOnly,
-                classification,
-                ministry,
-              }: Personnel) => ({
-                key: id,
-                status: status,
-
-                cells: [
-                  {
-                    key: uuidv4(),
-                    columnName: DashboardColumns.NAME,
-                    value: `${lastName.toUpperCase()},  ${firstName}`,
-                    className: tableClass(DashboardColumns.NAME, ''),
-                  },
-
-                  {
-                    key: uuidv4(),
-                    columnName: DashboardColumns.REGION,
-                    value: region,
-                    className: tableClass(
-                      DashboardColumns.REGION,
-                      region?.toLowerCase(),
-                    ),
-                  },
-                  {
-                    key: uuidv4(),
-                    columnName: DashboardColumns.LOCATION,
-                    value: workLocation,
-                    className: tableClass(
-                      DashboardColumns.LOCATION,
-                      workLocation?.toLowerCase(),
-                    ),
-                  },
-                  {
-                    key: uuidv4(),
-                    columnName: DashboardColumns.FUNCTION,
-                    value: `${experiences?.find((itm: any) => itm.functionName === filterValues.function)?.functionName}:${ExperienceName[experiences?.find((itm: any) => itm.functionName === filterValues.function)?.experienceType as keyof typeof ExperienceName]}`,
-                    className: experiences?.find(
-                      (itm: any) => itm.functionName === filterValues.function,
-                    )
-                      ? tableClass(DashboardColumns.FUNCTION, '')
-                      : 'hidden',
-                  },
-                  {
-                    key: uuidv4(),
-                    columnName: DashboardColumns.AVAILABILITY,
-                    value: filterValues.availabilityType
-                      ? getAvailabilityValue(filterValues.availabilityType)
-                      : AvailabilityTypeName.NOT_INDICATED,
-                    className: tableClass(
-                      DashboardColumns.AVAILABILITY,
-                      filterValues.availabilityType ??
-                        AvailabilityType.NOT_INDICATED,
-                    ),
-                  },
-                  {
-                    key: uuidv4(),
-                    columnName: DashboardColumns.TRAVEL,
-                    value: willingToTravel,
-                    className: tableClass(
-                      DashboardColumns.TRAVEL,
-                      willingToTravel ? 'yes' : 'no',
-                    ),
-                  },
-                  {
-                    key: uuidv4(),
-                    columnName: DashboardColumns.REMOTE,
-                    value: remoteOnly,
-                    className: tableClass(
-                      DashboardColumns.REMOTE,
-                      remoteOnly ? 'yes' : 'no',
-                    ),
-                  },
-                  {
-                    key: uuidv4(),
-                    columnName: DashboardColumns.UNION_MEMBERSHIP,
-                    value: classification,
-                    className: tableClass(
-                      DashboardColumns.UNION_MEMBERSHIP,
-                      classification,
-                    ),
-                  },
-                  {
-                    key: uuidv4(),
-                    columnName: DashboardColumns.MINISTRY,
-                    value: ministry,
-                    className: tableClass(
-                      DashboardColumns.MINISTRY,
-                      ministry?.toLowerCase(),
-                    ),
-                  },
-                ],
-              }),
-            ),
-          });
+        setTableData({
+          totalPages,
+          pageRange: truncatePageRange(totalPages, currentPage, pageRange),
+          totalRows: count,
+          rows: personnel.map(({ id, status, ...personnel }: Personnel) => ({
+            key: id,
+            status: status,
+            cells: renderCells(personnel, filterValues),
+          })),
+        });
       } catch (e) {
         handleError(e);
       }
@@ -321,7 +169,7 @@ export const useTable = () => {
     handleMultiSelect,
     handleSingleSelect,
     handleSearch,
-    showFunctionColumn,
+    showFunctionColumn: filterValues.function,
     filterValues,
     handleClose,
     handleCloseMany,
